Move node stub TS definitions
