+++ src/test-utils/mock-file-system.ts
@@ -1,4 +1,4 @@
-import type { FileSystem } from '../write-bin-template/types'
+import type { FileSystem } from '../../types/node-stub.d'
 import { mockify } from './mockify'
 
 // Feel free to fill in mkdirSync and rmSync if you think of something sensible
